3 - refactor toast service
