+++ src/components/Toast/Toast.tsx
@@ -12,15 +12,15 @@ import {Text} from '../Text/Text';
 const MAX_WIDTH = Dimensions.get('screen').width * 0.9;
 
 export function Toast() {
-  const {toast, hiddenToast} = useToast();
+  const {toast, hideToast} = useToast();
 
   useEffect(() => {
     if (toast) {
       setTimeout(() => {
-        hiddenToast();
+        hideToast();
       }, 2000);
     }
-  }, [hiddenToast, toast]);
+  }, [hideToast, toast]);
 
   if (!toast) {
     return null;
