+++ src/services/index.ts
@@ -1,1 +1,4 @@
+export * from './toast/toastTypes';
+export {ToastProvider} from './toast/Providers/ToastProvider';
+
 export * from './toast/useToast';
