+++ src/services/toast/Providers/ToastProvider.tsx
@@ -1,0 +1,27 @@
+import React, {createContext, useState} from 'react';
+
+import {ToastService, Toast} from '../toastTypes';
+
+export const ToastContext = createContext<ToastService>({
+  toast: null,
+  showToast: () => {},
+  hideToast: () => {},
+});
+
+export function ToastProvider({children}: React.PropsWithChildren<{}>) {
+  const [toast, setToast] = useState<ToastService['toast']>(null);
+
+  function showToast(_toast: Toast) {
+    setToast(_toast);
+  }
+
+  function hideToast() {
+    setToast(null);
+  }
+
+  return (
+    <ToastContext.Provider value={{toast, showToast, hideToast}}>
+      {children}
+    </ToastContext.Provider>
+  );
+}
