+++ src/services/toast/toastTypes.ts
@@ -1,0 +1,15 @@
+export interface Toast {
+  message: string;
+  type?: 'success' | 'error';
+  duration?: number;
+  action?: {
+    title: string;
+    onPress: () => void;
+  };
+}
+
+export interface ToastService {
+  toast: Toast | null;
+  showToast: (toast: Toast) => void;
+  hideToast: () => void;
+}
